+++ pennylane_lightning/_serialize.py
@@ -17,13 +17,24 @@ Helper functions for serializing quantum tapes.
 from typing import List, Tuple
 
 import numpy as np
-from pennylane import BasisState, Hadamard, Projector, QubitStateVector, Rot
+from pennylane import (
+    BasisState,
+    Hadamard,
+    Projector,
+    QubitStateVector,
+    Rot,
+)
 from pennylane.grouping import is_pauli_word
 from pennylane.operation import Observable, Tensor
 from pennylane.tape import QuantumTape
 
 try:
-    from .lightning_qubit_ops import StateVectorC128, ObsStructC128
+    from .lightning_qubit_ops import (
+        StateVectorC64,
+        ObsStructC64,
+        StateVectorC128,
+        ObsStructC128,
+    )
 except ImportError:
     pass
 
@@ -46,17 +57,25 @@ def _obs_has_kernel(obs: Observable) -> bool:
     return False
 
 
-def _serialize_obs(tape: QuantumTape, wires_map: dict) -> List:
+def _serialize_obs(tape: QuantumTape, wires_map: dict, use_csingle: bool = False) -> List:
     """Serializes the observables of an input tape.
 
     Args:
         tape (QuantumTape): the input quantum tape
         wires_map (dict): a dictionary mapping input wires to the device's backend wires
+        use_csingle (bool): whether to use np.complex64 instead of np.complex128
 
     Returns:
-        list(ObsStructC128): A list of observable objects compatible with the C++ backend
+        list(ObsStructC128 or ObsStructC64): A list of observable objects compatible with the C++ backend
     """
     obs = []
+
+    if use_csingle:
+        ctype = np.complex64
+        obs_py = ObsStructC64
+    else:
+        ctype = np.complex128
+        obs_py = ObsStructC128
 
     for o in tape.observables:
         is_tensor = isinstance(o, Tensor)
@@ -81,20 +100,20 @@ def _serialize_obs(tape: QuantumTape, wires_map: dict) -> List:
             if is_tensor:
                 for o_ in o.obs:
                     if not _obs_has_kernel(o_):
-                        params.append(o_.matrix.ravel().astype(np.complex128))
+                        params.append(o_.matrix.ravel().astype(ctype))
                     else:
                         params.append([])
             else:
-                params.append(o.matrix.ravel().astype(np.complex128))
+                params.append(o.matrix.ravel().astype(ctype))
 
-        ob = ObsStructC128(name, params, wires)
+        ob = obs_py(name, params, wires)
         obs.append(ob)
 
     return obs
 
 
 def _serialize_ops(
-    tape: QuantumTape, wires_map: dict
+    tape: QuantumTape, wires_map: dict, use_csingle: bool = False
 ) -> Tuple[List[List[str]], List[np.ndarray], List[List[int]], List[bool], List[np.ndarray]]:
     """Serializes the operations of an input tape.
 
@@ -103,6 +122,7 @@ def _serialize_ops(
     Args:
         tape (QuantumTape): the input quantum tape
         wires_map (dict): a dictionary mapping input wires to the device's backend wires
+        use_csingle (bool): whether to use np.complex64 instead of np.complex128
 
     Returns:
         Tuple[list, list, list, list, list]: A serialization of the operations, containing a list
@@ -114,6 +134,8 @@ def _serialize_ops(
     wires = []
     inverses = []
     mats = []
+
+    sv_py = StateVectorC64 if use_csingle else StateVectorC128
 
     uses_stateprep = False
 
@@ -132,7 +154,7 @@ def _serialize_ops(
             name = single_op.name if not is_inverse else single_op.name[:-4]
             names.append(name)
 
-            if getattr(StateVectorC128, name, None) is None:
+            if getattr(sv_py, name, None) is None:
                 params.append([])
                 mats.append(single_op.matrix)
 
@@ -145,4 +167,5 @@ def _serialize_ops(
             wires_list = single_op.wires.tolist()
             wires.append([wires_map[w] for w in wires_list])
             inverses.append(is_inverse)
+
     return (names, params, wires, inverses, mats), uses_stateprep